+++ src/components/PlayerView/index.tsx
@@ -14,7 +14,9 @@ import { useProxy } from '../../hooks/useProxy';
 import { FullscreenHandler, CustomMessageHandler } from '../../ui';
 import { FullscreenHandlerBridge } from '../../ui/fullscreenhandlerbridge';
 import { CustomMessageHandlerBridge } from '../../ui/custommessagehandlerbridge';
+import { ScalingMode } from '../../styleConfig';
 import { PictureInPictureConfig } from './pictureInPictureConfig';
+import { PlayerViewConfig } from './playerViewConfig';
 
 /**
  * Base `PlayerView` component props. Used to establish common
@@ -40,6 +42,12 @@ export interface BasePlayerViewProps {
    * Provides options to configure Picture in Picture playback.
    */
   pictureInPictureConfig?: PictureInPictureConfig;
+
+  /**
+   * Configures the visual presentation and behaviour of the `PlayerView`.
+   * The value must not be altered after setting it initially.
+   */
+  config?: PlayerViewConfig;
 }
 
 /**
@@ -62,6 +70,18 @@ export interface PlayerViewProps extends BasePlayerViewProps, PlayerViewEvents {
    * To use this property, a `FullscreenHandler` must be set.
    */
   isFullscreenRequested?: Boolean;
+
+  /**
+   * A value defining how the video is displayed within the parent container's bounds.
+   * Possible values are defined in `ScalingMode`.
+   */
+  scalingMode?: ScalingMode;
+
+  /**
+   * Can be set to `true` to request Picture in Picture mode, or `false` to request exit of Picture in Picture mode.
+   * Should not be used to get the current Picture in Picture state. Use `onPictureInPictureEnter` and `onPictureInPictureExit.
+   */
+  isPictureInPictureRequested?: Boolean;
 }
 
 /**
@@ -97,9 +117,12 @@ function dispatch(command: string, node: NodeHandle, ...args: any[]) {
 export function PlayerView({
   style,
   player,
+  config,
   fullscreenHandler,
   customMessageHandler,
   isFullscreenRequested = false,
+  scalingMode,
+  isPictureInPictureRequested = false,
   pictureInPictureConfig,
   ...props
 }: PlayerViewProps) {
@@ -179,6 +202,21 @@ export function PlayerView({
       dispatch('setFullscreen', node, isFullscreenRequested);
     }
   }, [isFullscreenRequested, nativeView]);
+
+  useEffect(() => {
+    const node = findNodeHandle(nativeView.current);
+    if (node) {
+      dispatch('setScalingMode', node, scalingMode);
+    }
+  }, [scalingMode, nativeView]);
+
+  useEffect(() => {
+    const node = findNodeHandle(nativeView.current);
+    if (node) {
+      dispatch('setPictureInPicture', node, isPictureInPictureRequested);
+    }
+  }, [isPictureInPictureRequested, nativeView]);
+
   return (
     <NativePlayerView
       ref={nativeView}
@@ -186,6 +224,7 @@ export function PlayerView({
       fullscreenBridge={fullscreenBridge.current}
       customMessageHandlerBridge={customMessageHandlerBridge.current}
       pictureInPictureConfig={pictureInPictureConfig}
+      config={config}
       onAdBreakFinished={proxy(props.onAdBreakFinished)}
       onAdBreakStarted={proxy(props.onAdBreakStarted)}
       onAdClicked={proxy(props.onAdClicked)}