+++ CHANGELOG.md
@@ -8,7 +8,10 @@
 - `PictureInPictureConfig` to `PlayerView` to allow configuring the Picture in Picture behavior
 - `PictureInPictureConfig.shouldEnterOnBackground` to start PiP automatically when the app transitions to background
 - `onPictureInPictureAvailabilityChanged` event is now emitted on iOS and tvOS in addition to Android
-- `BufferConfig` to control player buffer depth
+- `BufferConfig` to configure player buffer depth
+- `PlayerView.isPictureInPictureRequested` to programatically create a Picture in Picture request
+- `PlayerView.scalingMode` to allow changing the video scaling mode
+- `PlayerViewConfig` with a `UiConfig` to the `PlayerView` to enable configuration of the visual presentation and behaviour
 
 ### Changed